% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate Temperature Quantiles}
\usage{
calculate_quantiles(
  data_series,
  date_sequence,
  base.range,
  n,
  temp.qtiles,
  min.base.data.fraction.present
)
}
\arguments{
\item{data_series}{Numeric vector. Data series aligned with date sequence.}

\item{date_sequence}{Date vector. Complete date sequence.}

\item{base.range}{Numeric vector. Base range years.}

\item{n}{Integer. Window size for running averages.}

\item{temp.qtiles}{Numeric vector. Quantiles to calculate.}

\item{min.base.data.fraction.present}{Numeric. Minimum fraction of data required in base range.}
}
\value{
Data frame. Calculated quantiles.
}
\description{
Computes temperature quantiles based on data within a specified base range.
}
\keyword{internal}
