% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{calculate_standardized}
\alias{calculate_standardized}
\title{Calculate Standardized Indices}
\usage{
calculate_standardized(df, freq, base_range)
}
\arguments{
\item{df}{Data frame. Input data with Date and Value columns.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "seasonal".}

\item{base_range}{Numeric vector. Base range years.}
}
\value{
Data frame with standardized values.
}
\description{
Helper function to calculate standardized indices based on a reference period.
}
\keyword{internal}
