% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{calculate_wind_quantiles}
\alias{calculate_wind_quantiles}
\title{Calculate Wind Quantiles}
\usage{
calculate_wind_quantiles(
  wind_data_series,
  date_sequence,
  base_range,
  wind_qtile = 0.9
)
}
\arguments{
\item{wind_data_series}{Numeric vector. Wind data series aligned with date sequence.}

\item{date_sequence}{Date vector. Complete date sequence.}

\item{base_range}{Numeric vector. Base range years.}

\item{wind_qtile}{Numeric. Wind quantile to calculate (default is 0.90).}
}
\value{
Data frame. Calculated wind thresholds.
}
\description{
Computes the 90th percentile of wind speed for index calculations.
}
\keyword{internal}
