% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{cdd}
\alias{cdd}
\title{Calculate Consecutive Dry Days (CDD)}
\usage{
cdd(ci, spells_can_span_years = TRUE, monthly = TRUE)
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{spells_can_span_years}{Logical. Whether spells can span across years (default is TRUE).}

\item{monthly}{Logical. Whether to interpolate annual data to monthly data (default is TRUE).}
}
\value{
Data frame with dates and calculated CDD values.
}
\description{
Calculates the maximum length of consecutive dry days.
}
\examples{
\dontrun{
# Calculate annual CDD index
cdd_values <- cdd(ci, monthly = FALSE)
}
}
