% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{cdd_std}
\alias{cdd_std}
\title{Calculate Standardized CDD Index}
\usage{
cdd_std(ci, freq = c("monthly", "seasonal"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "seasonal".}
}
\value{
Data frame with dates and standardized CDD values.
}
\description{
Calculates the standardized consecutive dry days index.
}
\examples{
\dontrun{
# Calculate standardized CDD index on a monthly basis
cdd_std_values <- cdd_std(ci, freq = "monthly")
}
}
