% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{create_date_sequence}
\alias{create_date_sequence}
\title{Create Date Sequence}
\usage{
create_date_sequence(dates)
}
\arguments{
\item{dates}{Date vector. Original dates.}
}
\value{
Date vector. A complete sequence of dates.
}
\description{
Generates a complete sequence of dates from the earliest to the latest date provided.
}
\keyword{internal}
