% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{fill_data_series}
\alias{fill_data_series}
\title{Fill Data Series}
\usage{
fill_data_series(data, dates, date_sequence)
}
\arguments{
\item{data}{Numeric vector. Original data series.}

\item{dates}{Date vector. Dates corresponding to the data.}

\item{date_sequence}{Date vector. Complete date sequence.}
}
\value{
Numeric vector. Data series aligned with the date sequence.
}
\description{
Aligns the input data with the complete date sequence, filling missing dates with NA.
}
\keyword{internal}
