% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{get.series.lengths.at.ends}
\alias{get.series.lengths.at.ends}
\title{Get Series Lengths at Ends}
\usage{
get.series.lengths.at.ends(x, na.value = FALSE)
}
\arguments{
\item{x}{Logical vector.}

\item{na.value}{Logical. Value to assign to NA elements.}
}
\value{
Numeric vector. Lengths of runs at the ends.
}
\description{
Calculates the lengths of runs at the ends of a logical vector.
}
\keyword{internal}
