% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{get_series_lengths}
\alias{get_series_lengths}
\title{Get Series Lengths}
\usage{
get_series_lengths(bools)
}
\arguments{
\item{bools}{Logical vector.}
}
\value{
Numeric vector. Lengths of consecutive TRUE runs.
}
\description{
Calculates the lengths of consecutive TRUE values in a logical vector.
}
\keyword{internal}
