% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{monthly_to_seasonal}
\alias{monthly_to_seasonal}
\title{Convert Monthly Data to Seasonal Data}
\usage{
monthly_to_seasonal(data)
}
\arguments{
\item{data}{Data frame. Monthly data with Date and Value columns.}
}
\value{
Data frame with seasonal data.
}
\description{
Aggregates monthly data into seasonal averages.
}
\examples{
\dontrun{
# Assuming you have monthly data in a data frame 'monthly_data'
# with columns 'Date' (in 'YYYY-MM' format) and 'Value'
seasonal_data <- monthly_to_seasonal(monthly_data)
}
}
