% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{nday_consec_prec_max}
\alias{nday_consec_prec_max}
\title{Maximum Precipitation Over Consecutive Days}
\usage{
nday_consec_prec_max(
  daily_prec,
  date_factor,
  ndays,
  center_mean_on_last_day = FALSE
)
}
\arguments{
\item{daily_prec}{Numeric vector. Daily precipitation data.}

\item{date_factor}{Factor. Date grouping factor.}

\item{ndays}{Integer. Number of consecutive days.}

\item{center_mean_on_last_day}{Logical. Whether to center the mean on the last day.}
}
\value{
Numeric vector. Maximum precipitation amounts.
}
\description{
Calculates maximum precipitation over consecutive n days.
}
\keyword{internal}
