% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{running_mean}
\alias{running_mean}
\title{Running Mean}
\usage{
running_mean(vec, bin)
}
\arguments{
\item{vec}{Numeric vector. Input data.}

\item{bin}{Integer. Window size.}
}
\value{
Numeric vector. Running mean values.
}
\description{
Calculates the running mean of a vector over a specified window size.
}
\keyword{internal}
