% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{rx5day}
\alias{rx5day}
\title{Calculate Rx5day Index}
\usage{
rx5day(ci, freq = c("monthly", "annual"), center_mean_on_last_day = FALSE)
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}

\item{center_mean_on_last_day}{Logical. Whether to center the mean on the last day (default is FALSE).}
}
\value{
Data frame with dates and calculated Rx5day values.
}
\description{
Calculates the maximum consecutive 5-day precipitation amount.
}
\examples{
\dontrun{
# Calculate monthly Rx5day index
rx5day_values <- rx5day(ci, freq = "monthly")
}
}
