% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{t10p}
\alias{t10p}
\title{Calculate T10p Index}
\usage{
t10p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated T10p values.
}
\description{
Calculates the combined percentage of days when temperature is below the 10th percentile.
}
\examples{
\dontrun{
# Calculate monthly T10p index
t10p_values <- t10p(ci, freq = "monthly")
}
}
