% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{t10p_std}
\alias{t10p_std}
\title{Calculate Standardized T10p Index}
\usage{
t10p_std(ci, freq = c("monthly", "seasonal"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "seasonal".}
}
\value{
Data frame with dates and standardized T10p values.
}
\description{
Calculates the standardized T10p index.
}
\examples{
\dontrun{
Calculate standardized T10p index on a monthly basis
t10p_std_values <- t10p_std(ci, freq = "monthly")
}
}
