% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{t90p}
\alias{t90p}
\title{Calculate T90p Index}
\usage{
t90p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated T90p values.
}
\description{
Calculates the combined percentage of days when temperature is above the 90th percentile.
}
\examples{
\dontrun{
# Calculate monthly T90p index
t90p_values <- t90p(ci, freq = "monthly")
}
}
