% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{t90p_std}
\alias{t90p_std}
\title{Calculate Standardized T90p Index}
\usage{
t90p_std(ci, freq = c("monthly", "seasonal"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "seasonal".}
}
\value{
Data frame with dates and standardized T90p values.
}
\description{
Calculates the standardized T90p index.
}
\examples{
\dontrun{
# Calculate standardized T90p index on a monthly basis
t90p_std_values <- t90p_std(ci, freq = "monthly")
}
}
