% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{tapply.fast}
\alias{tapply.fast}
\title{Fast tapply Function}
\usage{
tapply.fast(X, INDEX, FUN = NULL, ..., simplify = TRUE)
}
\arguments{
\item{X}{Numeric vector. Data to be applied.}

\item{INDEX}{Factor. Indexing factor.}

\item{FUN}{Function. Function to apply.}

\item{simplify}{Logical. Whether to simplify the result.}
}
\value{
List or vector. Result of applying FUN to X grouped by INDEX.
}
\description{
Provides a faster implementation of tapply for factor-type indices.
}
\keyword{internal}
