% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{tn10p}
\alias{tn10p}
\title{Calculate TN10p Index}
\usage{
tn10p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated TN10p values.
}
\description{
Calculates the percentage of days when minimum temperature is below the 10th percentile.
}
\examples{
\dontrun{
# Calculate monthly TN10p index
tn10p_values <- tn10p(ci, freq = "monthly")
}
}
