% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{tn90p}
\alias{tn90p}
\title{Calculate TN90p Index}
\usage{
tn90p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated TN90p values.
}
\description{
Calculates the percentage of days when minimum temperature is above the 90th percentile.
}
\examples{
\dontrun{
# Calculate monthly TN90p index
tn90p_values <- tn90p(ci, freq = "monthly")
}
}
