% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{tx10p}
\alias{tx10p}
\title{Calculate TX10p Index}
\usage{
tx10p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated TX10p values.
}
\description{
Calculates the percentage of days when maximum temperature is below the 10th percentile.
}
\examples{
\dontrun{
# Calculate monthly TX10p index
tx10p_values <- tx10p(ci, freq = "monthly")
}
}
