% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{tx90p}
\alias{tx90p}
\title{Calculate TX90p Index}
\usage{
tx90p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated TX90p values.
}
\description{
Calculates the percentage of days when maximum temperature is above the 90th percentile.
}
\examples{
\dontrun{
# Calculate monthly TX90p index
tx90p_values <- tx90p(ci, freq = "monthly")
}
}
