% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{w90p}
\alias{w90p}
\title{Calculate W90p Index}
\usage{
w90p(ci, freq = c("monthly", "annual"))
}
\arguments{
\item{ci}{List. Climate input object created by \code{\link{climate_input}}.}

\item{freq}{Character. Frequency of calculation, either "monthly" or "annual".}
}
\value{
Data frame with dates and calculated W90p values.
}
\description{
Calculates the percentage of days when wind speed is above the 90th percentile.
}
\examples{
\dontrun{
# Calculate monthly W90p index
w90p_values <- w90p(ci, freq = "monthly")
}
}
