% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIACI.R
\name{w90p_std}
\alias{w90p_std}
\title{Calculate Standardized W90p Index}
\usage{
w90p_std(ci, freq = c("monthly", "seasonal"))
}
\arguments{
\item{ci}{List. Climate input object.}

\item{freq}{Character. Frequency of calculation, either "monthly", "seasonal".}
}
\value{
Data frame with dates and standardized W90p values.
}
\description{
Calculates the standardized W90p index.
}
\examples{
\dontrun{
# Calculate standardized W90p index on a monthly basis
w90p_std_values <- w90p_std(ci, freq = "monthly")
}
}
